Update CreateTestCaseResponse class. Change "Id" data type to "long".

+++ dev/src/TestManagement.API/TestManagement.API/Features/TestCases/Create/CreateTestCaseResponse.cs
@@ -2,7 +2,7 @@
 {
     public class CreateTestCaseResponse
     {
-        public int Id { get; set; } = 0;
+        public long Id { get; set; } = 0;
 
         public string Code { get; set; } = null!;
 
@@ -10,7 +10,7 @@
 
         public string Description { get; set; } = null!;
 
-        public int TestLevelId { get; set; } = 0;
+        public long TestLevelId { get; set; } = 0;
 
         public int VersionNumber { get; set; } = 0;
     }
